+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Now creates a new Custom Icon for Credential Types.
+- Now gathers Credential Types.
+- Now creates an `ATUsesType` edge between, Credentials and their Credential Type.
 - Now able to skip TLS/SSL verification using the `-k`/`skip-verify-ssl` flag.
 - Now gathers information on the target Ansible WorX/Tower instance using `/ping` before gathering resources.
 - Now gathers Credentials used by Job Templates and creates an `ATUses` edge between them.